After each GMRES solution the value stored in monitor.residual_norm() is now printed on the screen to see how much the GMRES solution is converged. For this purpose a new variable called solverNorm is introduced.
To understand how much time different parts of the code is using, timing functions and comments are added into the solve() and CUSPsolver() functions. They are commented out now.

CUSPsolver.cu is changed to use preconditioners with CUSP. But since, unexpectedly, they all slow down the solution, they are all commented out now. This should be investigated.

At the end of a converged run, a Tecplot file is now created even if the last iteration number is not a multiple of nDATiter.

cavityMeshGenerator.m is updated to print nDATiter and restart varibales to the INP files.