Many changes are done, but the most important one is about the storage of global matrices.
Column, Row and Value arrays [M], [K], [A] matrices are now stored only for one sub matrix. Same is true for the column and Row arrays of the [G] matrix. But Value array of the [G] matrix is divded into three and stored as three separate arrays.

Related to the above, some of MKL's matrix-vector-multiplication calls are also divided into three.

For NE=27,000 Cubical Cavity problem storage dropped from 3 GB to 2 GB, a considerable increase. Also run time is improved. For the 1st iterations step1() is now faster considerably.