.CPP and .CU files are renamed.
Many unnecessary copies between CPU and GPU are eliminated.

A small bug in convergence check in the GPU is corrected.

Same .CPP file now works both on the CPU and the GPU, depending on the definition of USECUDA compiler directive.

CubeMesh_NE125000 problem does not work on Tesla 2075 because its 6GB memory is not enough. Most of it is consumed by Z_chol data. For large problems Cholesky factorization seems to be problematic and PCG can be tried.