step2 can now use MKL_CG instead of Cholesky. First define USECUDA and a file with .Zcsr extension will be created. Stop the solver and recompile without USECUDA option. Z matrix will be read from the file and MKL_CG will be used in step2.
Although this works, it is complicated and should be simplified. We can calculate Z using Csparse as done before with the Cholesky solver. No file need to be created for the Z matrix in that way. But the use of CSparse library is also cumbersome. The problem here is sparse matrix-matrix multiplication in creating Z. Currently only CSparse and CUSP can do it. MKL can not. We should search for a simpler library to calculate it on the CPU.

Tested with CubeMesh_NE1000 problem.