A new file that uses PARDISO for Cholesky factorization on the CPU is added. PARDISO can run multi-threaded, CSparse can not. PARDISO can calculate the Cholesky factorization much faster than CSparse. There is no need to write the factorization to a file and read it from there for future runs. Also PARDISO's METIS based renumbering results in less number of nonzeros's for the factorized lower triangular matrix. CSparse is still necessary to caluclate Z itself (due to sparse matrix-matrix multiplication).
To be fair, Cholesky solver on the GPU should also use PARDISO's factorization. But it is not possible to extract any information from PARDISO. Alternative is CHOLMOD. CHOLMOD can also use METIS renumbering. One disadvantage of CHOLMOD is that it is not easy to install it on Windows. But Linux can use it w/o problems.