Fractional step solver that uses PCG in step 2 is updated. "_noCholesky" suffix is removed from the CPP and CU files. Previously these file names were used for the Cholesky version, but now they are used for the PCG version because this is the main solver that we'll be working on. Cholesky version may also be updated later.
There are several changes compared to the previous version.

No files related to the [Z] matrix are read/written. [Z] matrix is created in each run.

A new function named selectCUDAdevice() is generated and placed in CUDAcodes.cu file. It is called at the beginning of the main function.

initializeAndAllocateGPU() function is moved to CUDAcodes.cu file.

A new function named calculateMatrixA() is created. It is called from step1() function.

To have a cleaner screen output, PRINT_TIMES variable is introduced. When it is set to zero CPU time information inside the time loop is not printed on the screen.